migrated code to create gauge charts to a dedicated component
