+++ Frontend/src/app/dashboard/dashboard.component.scss
@@ -3,16 +3,4 @@
     padding-bottom: 150px;
 }
 
-.co2-container, .humidity-container, .temperature-container{
-    height: 200px;
-    width: 250px;
-    margin-right: auto;
-    margin-left: auto;
-    margin-top: 20px;
-    margin-bottom: 20px;
-}
 
-#co2Chart, #humidityChart, #temperatureChart{
-    margin-right: auto;
-    margin-left: auto;
-}
