+++ Frontend/src/app/gauge/gauge.component.html
@@ -1,1 +1,3 @@
-<p>gauge works!</p>
+<div class="gauge-container">
+    <canvas [id]="gauge.id"></canvas>
+</div>
