+++ Frontend/src/app/gauge/gauge.component.scss
@@ -1,0 +1,14 @@
+.gauge-container {
+    height: 200px;
+    width: 250px;
+    margin-right: auto;
+    margin-left: auto;
+    margin-top: 20px;
+    margin-bottom: 20px;
+}
+
+canvas {
+    margin-right: auto;
+    margin-left: auto;
+
+}
