+++ Frontend/src/app/gauge/gauge.model.ts
@@ -1,0 +1,3 @@
+export class Gauge {
+    constructor(public id: string, public color: string){}
+}
